refactor: 출력 위한 결과를 반환하게 수정
- PlayerResult와 ParticipantResult를 만들어 결과를 Controller로 반환했다.
