+++ src/main/kotlin/domain/result/ParticipantsResult.kt
@@ -1,0 +1,5 @@
+package domain.result
+
+import domain.participants.Dealer
+
+class ParticipantsResult(val dealer: Dealer, val playerResult: List<PlayerResult>)
