+++ src/main/kotlin/domain/result/PlayerResult.kt
@@ -1,0 +1,5 @@
+package domain.result
+
+import domain.participants.Player
+
+class PlayerResult(val player: Player, val result: Result)
