+++ src/test/kotlin/domain/result/ResultTest.kt
@@ -1,4 +1,4 @@
-package domain.judge
+package domain.result
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
